Dodanie uzupełnienie leczenia o HealthPotionLarge, dodanie tekstury do HPL
