+++ src/main/java/logic/collectibles/Collectible.java
@@ -1,5 +1,6 @@
 package logic.collectibles;
 
+import graphics.Engine;
 import logic.Entity;
 
 public class Collectible extends Entity {
@@ -10,6 +11,19 @@ public class Collectible extends Entity {
 
     public Collectible(float posX, float posY, float width, float height, String textureName) {
         super(posX, posY, width, height, textureName);
+    }
+
+    @Override
+    public void init() {
+        super.init();
+        rectangle.initGL(this.textureName, "rectangle.vert.glsl", "rectangle.frag");
+    }
+
+    @Override
+    public void update() {
+        if (Engine.gameplay.KORKOWY.getRectangle().collidesWith(this.getRectangle())) {
+            usage();
+        }
     }
 
     public void usage() {}
